test(media-engine): cover 86F parsing and reconstruction

+++ src/GWGUI.MediaEngine/Containers/I86f/I86fExceptions.cs
@@ -14,5 +14,5 @@ internal static class I86fExceptions
     /// <summary>Signale un nombre de cellules de bits nul ou négatif.</summary>
     public static InvalidDataException InvalidBitCount(int logicalTrack, int bitCount) => new($"86F track {logicalTrack} has an invalid bit-cell count of {bitCount}.");
     /// <summary>Signale une charge utile de piste tronquée.</summary>
-    public static InvalidDataException TruncatedTrack(int logicalTrack, int offset, int nextOffset, int expectedLength, int availableLength) => new($"86F track {logicalTrack} is truncated between offsets {offset} and {nextOffset}: {expectedLength} data bytes are required, {availableLength} are available.");
+    public static InvalidDataException TruncatedTrack(int logicalTrack, int offset, int nextOffset, long expectedLength, int availableLength) => new($"86F track {logicalTrack} is truncated between offsets {offset} and {nextOffset}: {expectedLength} data bytes are required, {availableLength} are available.");
 }
